cash_register/requirements.txt
Django>=5.1
djangorestframework>=3.14.0
pdfkit>=1.0.0
qrcode>=7.3.1
psycopg2